+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - URL support for CSS properties
 - Build config option for copy assets folder
 - Build config option for static assets folder
+- Build config option for additional Webpack rules
+- Build config option for additional Webpack plugins
 
 ## [1.0.0-rc.13] - 28.08.2021
 ### Added